[241127] feat: SolplaceLogsNewPage 구현
